[Update]Milk 클래스 추가하면서 Beverage를 부모클래스로 활용
